+++ example/jest.config.js
@@ -6,8 +6,8 @@ module.exports = {
     '!src/__tests__/**/*.js',
     '!src/__mocks__/**/*.js',
   ],
-  setupFiles: ['./jest-setup.js'],
-  setupTestFrameworkScriptFile: './node_modules/jest-enzyme/lib/index.js',
+  setupFiles: ['./jest.setup.js'],
+  setupFilesAfterEnv: ['./node_modules/jest-enzyme/lib/index.js'],
   snapshotSerializers: ['enzyme-to-json/serializer'],
   testPathIgnorePatterns: ['<rootDir>/(dist|node_modules)/'],
 }